Move file related functions
